% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_loglik_am.R
\name{neg_loglik_am}
\alias{neg_loglik_am}
\title{Maximum likelihood estimator (MLE) for right-censored covariates in normal linear regression}
\usage{
neg_loglik_am(params, Y, X, C, Z = NULL, data)
}
\arguments{
\item{params}{Model parameter values.}

\item{Y}{Name of column variable.}

\item{X}{Name of censored covariate variable.}

\item{C}{Name of observed version of \code{X}.}

\item{Z}{(Optional) name(s) of additional fully observed covariates. Default is \code{Z = NULL}}

\item{data}{A dataframe containing at least columns \code{Y}, \code{X}, \code{C}, \code{Z}.}
}
\value{
A list with the following two elements:
\item{coeff}{A dataframe containing coefficient and standard error estimates at convergence.}
\item{code}{An integer indicating why the optimization process terminated. See \code{?nlm} for details on values.}
}
\description{
Maximum likelihood estimator (MLE) for right-censored covariates in normal linear regression
}
